+++ NEWS.md
@@ -1,3 +1,7 @@
-# mrf2d 0.1.0 (2019-07-25)
+# mrf2d 0.3
 
-* Testing news.
+  * Co-ocurrence histogram computing is now exported in the `cohist()` function.
+  * Computing of conditional probabilities for individual positions are now exported in the `cp_mrf2d()` function.
+  * All the conversions from/to arrays to/from vectors are now available. See the help file for `smr_stat`.
+  * Additional example data included in `field1` and `hfield1`.
+  * Documentation improved, including paper available in: https://arxiv.org/abs/2006.00383